Sorted the rendering buffer

+++ Sly1/gl.h
@@ -25,8 +25,8 @@ class GL
 		// Screen Object
 		GLuint sbo;
 
-		int width  = 800;
-		int height = 800;
+		float width  = 800;
+		float height = 800;
 
 		void InitGL();
 		void TerminateGL();
